Structure, Input and Level states
Changed the structure of inheritance leaving only Initializable so that controllers only subscribe to level states they need eg. pause, resume, load, etc.

Refactored the Input controller to broadcast events on Update so that UI Input is processed first.

Level states are now changed by calling static methods in the LevelController class which calls the appropriate events. The OnResume event is fired in LateUpdate, whilst the OnPause event is fired when the pause button is pressed.